rename fetchActorTopMovie to fetchActorMovie
