+++ data/repositories/vod/src/main/java/com/sanaa/vod/dataSource/remote/actor/RemoteActorDataSource.kt
@@ -10,7 +10,7 @@ interface RemoteActorDataSource {
 
     suspend fun getActorImages(actorId: Int): List<ImageDto>
 
-    suspend fun getActorTopMovies(actorId: Int): List<ActorCastCreditDto>
+    suspend fun getActorMovies(actorId: Int): List<ActorCastCreditDto>
 
-    suspend fun getActorTopTvShows(actorId: Int): List<ActorCastCreditDto>
+    suspend fun getActorTvShows(actorId: Int): List<ActorCastCreditDto>
 }
