+++ data/repositories/vod/src/main/java/com/sanaa/vod/repository/ActorRepositoryImpl.kt
@@ -40,7 +40,7 @@ class ActorRepositoryImpl(
 
     override suspend fun getActorTopMovies(id: Int): List<Movie> =
         safeCall("Failed to retrieve top movies for actor ID: $id") {
-            remoteDataSource.getActorTopMovies(id).map {
+            remoteDataSource.getActorMovies(id).map {
                 it.toMovie()
             }.take(20)
         }
@@ -48,7 +48,7 @@ class ActorRepositoryImpl(
 
     override suspend fun getActorTopTvShows(id: Int): List<TvSeries> =
         safeCall("Failed to retrieve top TV series for actor ID: $id") {
-            remoteDataSource.getActorTopTvShows(id).map {
+            remoteDataSource.getActorTvShows(id).map {
                 it.toTvSeries()
             }.take(20)
         }
